refactor(view): trim board_sort.rs under 300 lines, tighten dep-lock docs
Addresses two findings from a workflow-backed high-effort review of #546:

- model/board_sort.rs landed at 304 lines, four over the guideline this
  split existed to satisfy. test_board_sort_field_config_string_is_canonical
  never touched Model at all -- it's a pure round-trip test of
  BoardSortField's Display/FromStr, only incidentally placed here.
  Relocated to model/boards.rs (194 -> 216 lines), which has headroom.
  board_sort.rs is now 282 lines.
- CLAUDE.md's kanban-view description said "the compiler is the
  guarantee" for the dropped deps-allowlist script, which overclaims:
  the compiler only stops the crate from *using* an undeclared
  dependency, not from a future PR *declaring* one -- that's caught by
  review, not the type system. Tightened CLAUDE.md's wording to say so
  precisely. lib.rs's doc comment and kanban-view/README.md's mention
  carried the same over-claim; simplified both back to a plain
  statement of fact instead, since crate/README docs aren't the place
  for that process reasoning -- CLAUDE.md already is.

The review's third finding (a script deleted, replaced by inaccurate
docs) was reviewed and declined: the script's removal itself was a
deliberate, discussed decision (Cargo.toml already is the allowlist,
fully visible in any diff; the automated reviewer had no visibility
into that conversation), not an accidental regression. Only the
docs needed fixing, not the removal itself.

The fourth finding (make_card test helper duplicated between model/mod.rs
and model/cards.rs) was reviewed and left as-is, matching this crate's
already-established convention of accepting small test-fixture
duplication across sibling test modules rather than adding shared
test-support plumbing for a 3-line helper.
