+++ crates/kanban-view/src/lib.rs
@@ -2,8 +2,8 @@
 //!
 //! Sits below `kanban-tui`/`kanban-web` and above `kanban-domain`/`kanban-core`
 //! in the dependency graph. Deliberately does not depend on `kanban-service`
-//! or any rendering framework — this `Cargo.toml` simply never declares one,
-//! so the compiler is the guarantee; each consumer brings its own.
+//! or any rendering framework — this `Cargo.toml` simply never declares one.
+//! Each consumer (`kanban-tui`, `kanban-web`) brings its own rendering stack.
 
 pub mod card_list;
 pub mod card_list_component;
