+++ crates/kanban-view/src/model/boards.rs
@@ -72,6 +72,28 @@ mod tests {
     use kanban_domain::{Board, Snapshot};
 
     #[test]
+    fn test_board_sort_field_config_string_is_canonical() {
+        // The on-disk board_sort_field string is the domain `Display` spelling,
+        // and it round-trips back through the domain `FromStr` — one canonical
+        // spelling, no TUI-local PascalCase divergence (KAN-955).
+        use std::str::FromStr;
+        for field in [
+            BoardSortField::Position,
+            BoardSortField::Name,
+            BoardSortField::CreatedAt,
+            BoardSortField::ArchivedAt,
+        ] {
+            let s = field.to_string();
+            assert_eq!(
+                BoardSortField::from_str(&s),
+                Ok(field),
+                "config string {s:?} must round-trip through the domain FromStr"
+            );
+        }
+        assert_eq!(BoardSortField::ArchivedAt.to_string(), "archived_at");
+    }
+
+    #[test]
     fn test_displayed_boards_partition_cached_on_load() {
         use kanban_domain::Archived;
         let mut m = Model::default();
